[BACKLOG-92] Adds the plugin init code for the SqlContext. Refactors it into a super class called Context.
