+++ src/main/java/org/pentaho/mondrian/tck/MondrianContext.java
@@ -1,49 +1,26 @@
 package org.pentaho.mondrian.tck;
 
 
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
-import mondrian.olap.MondrianProperties;
-import mondrian.rolap.RolapUtil;
-import org.olap4j.CellSet;
-import org.olap4j.OlapConnection;
-import org.olap4j.OlapException;
-import org.olap4j.OlapStatement;
-
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Properties;
 import java.util.concurrent.ExecutionException;
 
-public class MondrianContext {
+import mondrian.olap.MondrianProperties;
+import mondrian.rolap.RolapUtil;
 
-  public static final Properties testProperties;
+import org.olap4j.CellSet;
+import org.olap4j.OlapConnection;
+import org.olap4j.OlapException;
+import org.olap4j.OlapStatement;
 
-  static {
-    try {
-      testProperties = loadTestProperties();
-      if ( Boolean.parseBoolean( testProperties.getProperty( "register.big-data-plugin" ) ) ) {
-        BigDataPluginUtil.prepareBigDataPlugin(
-          new File( testProperties.getProperty( "big-data-plugin.folder" ) ),
-          testProperties.getProperty( "active.hadoop.configuration" ) );
-      }
-    } catch ( Exception e ) {
-      throw new RuntimeException( e );
-    }
-  }
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.CacheLoader;
+import com.google.common.cache.LoadingCache;
 
-  private static Properties loadTestProperties() throws IOException {
-    Properties testProperties = new Properties();
-    testProperties.load( new BufferedReader( new FileReader( "test.properties" ) ) );
-    return testProperties;
-  }
+public class MondrianContext extends Context {
 
   private static final LoadingCache<String, MondrianContext> instances =
     CacheBuilder.newBuilder().build(new CacheLoader<String, MondrianContext>() {
